refactor: move reptree-independent functions into a util folder; comments
