+++ src/index.ts
@@ -13,7 +13,7 @@ export * from './treeTypes';
 export * from './operations';
 
 // Utilities
-export { default as uuid } from './uuid';
+export { default as uuid } from './utils/uuid';
 
 // Reactive helpers (opt-in)
 export { bindVertex } from './reactive';
